misc/web-abat-tests/abat/common/websocket/websocket_WebSocket_close_3000.html

<!DOCTYPE html>
<!--
Copyright (c) 2012 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Liu, Shilong <shilongx.liu@intel.com>
        Li, HaoX <haox.li@intel.com>

-->

<html>
  <head>
    <title>WebSocket Test: websocket_WebSocket_close_3000</title>
    <link rel="author" title="Intel" href="http://www.intel.com" />
    <link rel="help" href="http://www.w3.org/TR/2012/CR-websockets-20120920/#dom-websocket-close" />
    <meta name="flags" content="" />
    <meta name="assert" content="Check if the value of the argument code of WebSocket.close is 3000 that the websocket can be close normally" />
    <script type="text/javascript" src="../resources/testharness.js"></script>
    <script type="text/javascript" src="../resources/testharnessreport.js"></script>
    <script type="text/javascript" src="support/websocket.js"></script>
  </head>
  <body>
    <div id="log"></div>
    <script type="text/javascript">
        var t = async_test(document.title);
        var webSocket = new CreateWebSocket();
        var isOpenCalled = false;
        webSocket.addEventListener('open', function (evt) {
            webSocket.close(3000);
            isOpenCalled = true;
        }, true);

        webSocket.addEventListener('close', t.step_func(function (evt) {
            assert_true(isOpenCalled, "WebSocket connection should be opened");
            assert_equals(webSocket.readyState, webSocket.CLOSED, "readyState should be 3(CLOSED)");
            assert_equals(evt.wasClean, true, "wasClean should be TRUE");
            t.done();
        }), true);
    </script>
  </body>
</html>
